+++ src/frameworks.test.ts
@@ -1,7 +1,7 @@
 import { Counter, makeGraph, runGraph } from "./util/dependencyGraph";
 import { expect, test, vi } from "vitest";
 import { FrameworkInfo, TestConfig } from "./util/frameworkTypes";
-import { frameworkInfo } from "./config";
+import { frameworkInfo } from "./configFrameworks";
 
 (await Promise.all(frameworkInfo.map((frameworkLoader) => frameworkLoader()))).forEach((frameworkInfo) => frameworkTests(frameworkInfo));
