+++ src/kairoBench.ts
@@ -8,6 +8,7 @@ import { triangle } from "./kairo/triangle";
 import { unstable } from "./kairo/unstable";
 import { fastestTest } from "./util/benchRepeat";
 import { ReactiveFramework } from "./util/reactiveFramework";
+import { PerfResultCallback } from "./util/perfResult";
 
 const cases = [
   avoidablePropagation,
@@ -20,7 +21,7 @@ const cases = [
   unstable,
 ];
 
-export async function kairoBench(framework: ReactiveFramework) {
+export async function kairoBench(framework: ReactiveFramework, resultCallback: PerfResultCallback) {
   for (const c of cases) {
     const iter = framework.withBuild(() => {
       const iter = c(framework);
@@ -36,7 +37,7 @@ export async function kairoBench(framework: ReactiveFramework) {
       }
     });
 
-    process.send?.({
+    resultCallback({
       framework: framework.name,
       test: c.name,
       time: timing.time,
