+++ src/util/perfResult.ts
@@ -1,0 +1,7 @@
+export interface PerfResult {
+  framework: string;
+  test: string;
+  time: number;
+}
+
+export type PerfResultCallback = (result: PerfResult) => void;
