+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "outDir": "dist",
     "strict": true,
     "target": "ESNext",
     "moduleResolution": "Bundler",
